disable users #2419 #4475
This commit adds an API endpoint for disabling users.
Once users are disabled:

- They lose all their roles.
- They lose pending file access requests.
- They lose group memberships.
- They lose banner messages.
- They lose all notifications.
- They get an error on login. (Builtin/Shib/OAuth tested.)
- They get an error if they try to use the API.
- They cannot be assigned roles.
- They cannot be added to groups.
- They cannot initiate password reset.
- They cannot confirm their email (from a previous link).

+++ src/main/java/edu/harvard/iq/dataverse/confirmemail/ConfirmEmailServiceBean.java
@@ -169,6 +169,10 @@ public class ConfirmEmailServiceBean {
                 long nowInMilliseconds = new Date().getTime();
                 Timestamp emailConfirmed = new Timestamp(nowInMilliseconds);
                 AuthenticatedUser authenticatedUser = confirmEmailData.getAuthenticatedUser();
+                if (authenticatedUser.isDisabled()) {
+                    logger.fine("User is disabled.");
+                    return null;
+                }
                 authenticatedUser.setEmailConfirmed(emailConfirmed);
                 em.remove(confirmEmailData);
                 return goodTokenCanProceed;
